chore: create two packages to allocate the dto with distinct responsibilities request x response and move the dto classes to the new packages.
