+++ src/main/java/github/io/api_voting_challenge/controller/AgendaController.java
@@ -1,8 +1,8 @@
 package github.io.api_voting_challenge.controller;
 
 import github.io.api_voting_challenge.controller.docs.AgendaControllerDocs;
-import github.io.api_voting_challenge.dto.AgendaRequest;
-import github.io.api_voting_challenge.dto.AgendaResponse;
+import github.io.api_voting_challenge.dto.request.AgendaRequest;
+import github.io.api_voting_challenge.dto.response.AgendaResponse;
 import github.io.api_voting_challenge.service.AgendaService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
