+++ src/main/java/github/io/api_voting_challenge/dto/response/VoteResultResponse.java
@@ -1,4 +1,4 @@
-package github.io.api_voting_challenge.dto;
+package github.io.api_voting_challenge.dto.response;
 
 import lombok.Builder;
