+++ src/main/java/github/io/api_voting_challenge/dto/response/VotingSessionResponse.java
@@ -1,4 +1,4 @@
-package github.io.api_voting_challenge.dto;
+package github.io.api_voting_challenge.dto.response;
 
 import github.io.api_voting_challenge.model.enums.Status;
 import lombok.Builder;
