+++ src/main/java/github/io/api_voting_challenge/mapper/AgendaMapper.java
@@ -1,7 +1,7 @@
 package github.io.api_voting_challenge.mapper;
 
-import github.io.api_voting_challenge.dto.AgendaRequest;
-import github.io.api_voting_challenge.dto.AgendaResponse;
+import github.io.api_voting_challenge.dto.request.AgendaRequest;
+import github.io.api_voting_challenge.dto.response.AgendaResponse;
 import github.io.api_voting_challenge.model.Agenda;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
