+++ src/main/java/github/io/api_voting_challenge/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package github.io.api_voting_challenge.mapper;
 
-import github.io.api_voting_challenge.dto.UserRequest;
-import github.io.api_voting_challenge.dto.UserResponse;
+import github.io.api_voting_challenge.dto.request.UserRequest;
+import github.io.api_voting_challenge.dto.response.UserResponse;
 import github.io.api_voting_challenge.model.User;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
