+++ src/main/java/github/io/api_voting_challenge/service/AgendaService.java
@@ -1,7 +1,7 @@
 package github.io.api_voting_challenge.service;
 
-import github.io.api_voting_challenge.dto.AgendaRequest;
-import github.io.api_voting_challenge.dto.AgendaResponse;
+import github.io.api_voting_challenge.dto.request.AgendaRequest;
+import github.io.api_voting_challenge.dto.response.AgendaResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
