+++ src/main/java/github/io/api_voting_challenge/service/UserService.java
@@ -1,7 +1,7 @@
 package github.io.api_voting_challenge.service;
 
-import github.io.api_voting_challenge.dto.UserRequest;
-import github.io.api_voting_challenge.dto.UserResponse;
+import github.io.api_voting_challenge.dto.request.UserRequest;
+import github.io.api_voting_challenge.dto.response.UserResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
