+++ src/main/java/github/io/api_voting_challenge/service/VoteService.java
@@ -1,7 +1,7 @@
 package github.io.api_voting_challenge.service;
 
-import github.io.api_voting_challenge.dto.VoteRequest;
-import github.io.api_voting_challenge.dto.VoteResultResponse;
+import github.io.api_voting_challenge.dto.request.VoteRequest;
+import github.io.api_voting_challenge.dto.response.VoteResultResponse;
 
 public interface VoteService {
     void registerVote(Long sessionId, VoteRequest voteRequest);
