+++ src/main/java/github/io/api_voting_challenge/service/VotingSessionScheduler.java
@@ -1,6 +1,6 @@
 package github.io.api_voting_challenge.service;
 
-import github.io.api_voting_challenge.dto.VotingSessionResponse;
+import github.io.api_voting_challenge.dto.response.VotingSessionResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
