+++ src/main/java/github/io/api_voting_challenge/service/VotingSessionService.java
@@ -1,7 +1,7 @@
 package github.io.api_voting_challenge.service;
 
-import github.io.api_voting_challenge.dto.VotingSessionRequest;
-import github.io.api_voting_challenge.dto.VotingSessionResponse;
+import github.io.api_voting_challenge.dto.request.VotingSessionRequest;
+import github.io.api_voting_challenge.dto.response.VotingSessionResponse;
 
 public interface VotingSessionService {
     VotingSessionResponse openVotingSession(VotingSessionRequest votingSessionRequest);
