+++ src/test/java/github/io/api_voting_challenge/fixtures/UserFixtures.java
@@ -1,8 +1,8 @@
 package github.io.api_voting_challenge.fixtures;
 
 
-import github.io.api_voting_challenge.dto.UserRequest;
-import github.io.api_voting_challenge.dto.UserResponse;
+import github.io.api_voting_challenge.dto.request.UserRequest;
+import github.io.api_voting_challenge.dto.response.UserResponse;
 import github.io.api_voting_challenge.model.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
